new pi;uint_16 change to float

+++ device/device.h
@@ -422,13 +422,13 @@ extern struct pi_t currentLoopPI;		//do not initialize; configure
 extern struct pid_t pid_tArray[5];  	//do not initialize; configure
 extern float posIntegralArray[5];		//need initialize to 0
 extern float currIntegralArray[10];    //need initialize to 0
-extern uint16_t rotorPosition[5];		//need initialize to 0
-extern uint16_t coilCurrent[10];		//need initialize to 0
+extern float rotorPosition[5];		//need initialize to 0
+extern float coilCurrent[10];		//need initialize to 0
 extern uint16_t pwmDuty[10];			//need initialize to 2500
-extern uint16_t forwardFirstPos[5];		//need initialize to 0
-extern uint16_t refPosition[5];			//need configure
-extern uint16_t refCurrent[10];			//need initialize to 3750
-extern uint16_t coilBiasCurrent[5];		//do not initialize; configure
+extern float forwardFirstPos[5];		//need initialize to 0
+extern float refPosition[5];			//need configure
+extern float refCurrent[10];			//need initialize to 3750
+extern float coilBiasCurrent[5];		//do not initialize; configure
 extern uint32_t rawPosData[5];
 extern uint32_t rawCurrData[10];
